Add py3 suffix to tests
